+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+import 'dart:ui';
+
 import 'package:flutter/material.dart';
 import 'package:pin_input_text_field/pin_input_text_field.dart';
 
@@ -76,16 +78,14 @@ class _MyHomePageState extends State<MyHomePage> {
       GlobalKey<ScaffoldState>(debugLabel: 'home page global key');
 
   /// Decorate the outside of the Pin.
-  PinDecoration _pinDecoration = UnderlineDecoration(
-    enteredColor: Colors.green,
-    hintText: _kDefaultHint,
-  );
+  PinDecoration _pinDecoration;
 
   /// Control whether show the obscureCode.
   bool _obscureEnable = false;
 
   PinEntryType _pinEntryType = PinEntryType.underline;
-  Color _solidColor = Colors.black12;
+  ColorBuilder _solidColor =
+      PinListenColorBuilder(Colors.grey, Colors.grey[400]);
   bool _solidEnable = false;
 
   /// Control whether textField is enable.
@@ -108,6 +108,7 @@ class _MyHomePageState extends State<MyHomePage> {
       debugPrint('controller execute. pin:${_pinEditingController.text}');
     });
     super.initState();
+    _selectedMenu(PinEntryType.underline);
   }
 
   @override
@@ -122,8 +123,8 @@ class _MyHomePageState extends State<MyHomePage> {
       case PinEntryType.underline:
         setState(() {
           _pinDecoration = UnderlineDecoration(
-            enteredColor: Colors.green,
-            solidColor: _solidEnable ? _solidColor : null,
+            colorBuilder: PinListenColorBuilder(Colors.cyan, Colors.green),
+            bgColorBuilder: _solidEnable ? _solidColor : null,
             obscureStyle: ObscureStyle(
               isTextObscure: _obscureEnable,
               obscureText: '😂',
@@ -135,7 +136,7 @@ class _MyHomePageState extends State<MyHomePage> {
       case PinEntryType.boxTight:
         setState(() {
           _pinDecoration = BoxTightDecoration(
-            solidColor: _solidEnable ? _solidColor : null,
+            bgColorBuilder: _solidEnable ? _solidColor : null,
             obscureStyle: ObscureStyle(
               isTextObscure: _obscureEnable,
               obscureText: '👿',
@@ -147,8 +148,9 @@ class _MyHomePageState extends State<MyHomePage> {
       case PinEntryType.boxLoose:
         setState(() {
           _pinDecoration = BoxLooseDecoration(
-            enteredColor: Colors.green,
-            solidColor: _solidEnable ? _solidColor : null,
+            strokeColorBuilder:
+                PinListenColorBuilder(Colors.cyan, Colors.green),
+            bgColorBuilder: _solidEnable ? _solidColor : null,
             obscureStyle: ObscureStyle(
               isTextObscure: _obscureEnable,
               obscureText: '☺️',
@@ -160,10 +162,9 @@ class _MyHomePageState extends State<MyHomePage> {
       case PinEntryType.circle:
         setState(() {
           _pinDecoration = CirclePinDecoration(
-            enteredColor: Colors.green,
-            solidColor: _solidEnable ? _solidColor : null,
-            strokeColor: Colors.black,
-            strokeWidth: 4,
+            bgColorBuilder: _solidEnable ? _solidColor : null,
+            strokeColorBuilder:
+                PinListenColorBuilder(Colors.cyan, Colors.green),
             obscureStyle: ObscureStyle(
               isTextObscure: _obscureEnable,
               obscureText: '🤪',
@@ -279,7 +280,7 @@ class _MyHomePageState extends State<MyHomePage> {
               onChanged: (pin) {
                 debugPrint('onChanged execute. pin:$pin');
               },
-              enableInteractiveSelection: true,
+              enableInteractiveSelection: false,
             ),
           ),
         ],
@@ -413,6 +414,7 @@ class ExampleDecoration extends PinDecoration {
     TextStyle errorTextStyle,
     String hintText,
     TextStyle hintTextStyle,
+    ColorBuilder bgColorBuilder,
   }) : super(
           textStyle: textStyle,
           obscureStyle: obscureStyle,
@@ -430,24 +432,27 @@ class ExampleDecoration extends PinDecoration {
     TextStyle errorTextStyle,
     String hintText,
     TextStyle hintTextStyle,
-    Color solidColor,
+    ColorBuilder bgColorBuilder,
   }) {
     return ExampleDecoration(
-      textStyle: textStyle ?? this.textStyle,
-      obscureStyle: obscureStyle ?? this.obscureStyle,
-      errorText: errorText ?? this.errorText,
-      errorTextStyle: errorTextStyle ?? this.errorTextStyle,
-      hintText: hintText ?? this.hintText,
-      hintTextStyle: hintTextStyle ?? this.hintTextStyle,
-    );
+        textStyle: textStyle ?? this.textStyle,
+        obscureStyle: obscureStyle ?? this.obscureStyle,
+        errorText: errorText ?? this.errorText,
+        errorTextStyle: errorTextStyle ?? this.errorTextStyle,
+        hintText: hintText ?? this.hintText,
+        hintTextStyle: hintTextStyle ?? this.hintTextStyle,
+        bgColorBuilder: bgColorBuilder);
   }
+
+  @override
+  void notifyChange(String pin) {}
 
   @override
   void drawPin(
     Canvas canvas,
     Size size,
     String text,
-    pinLength,
+    int pinLength,
     ThemeData themeData,
   ) {
     /// You can draw anything you want here.