+++ lib/src/decoration/decoration_circle.dart
@@ -10,14 +10,11 @@ class CirclePinDecoration extends PinDecoration implements SupportGap {
   /// The gaps between every two adjacent box, higher priority than [gapSpace].
   final List<double> gapSpaces;
 
-  /// The box border color.
-  final Color strokeColor;
+  /// The box border color of index character.
+  final ColorBuilder strokeColorBuilder;
 
-  /// The box inside solid color, sometimes it equals to the box background.
-  final Color solidColor;
-
-  /// The border changed color when user enter pin.
-  final Color enteredColor;
+  // The background color of index character.
+  final ColorBuilder bgColorBuilder;
 
   const CirclePinDecoration({
     TextStyle textStyle,
@@ -28,18 +25,18 @@ class CirclePinDecoration extends PinDecoration implements SupportGap {
     TextStyle hintTextStyle,
     this.gapSpace: 16,
     this.gapSpaces,
-    this.strokeColor: Colors.cyan,
+    @required this.strokeColorBuilder,
     this.strokeWidth: 1,
-    this.solidColor,
-    this.enteredColor,
-  }) : super(
+    this.bgColorBuilder,
+  })  : assert(strokeColorBuilder != null),
+        super(
           textStyle: textStyle,
           obscureStyle: obscureStyle,
           errorText: errorText,
           errorTextStyle: errorTextStyle,
           hintText: hintText,
           hintTextStyle: hintTextStyle,
-          solidColor: solidColor,
+          bgColorBuilder: bgColorBuilder,
         );
 
   @override
@@ -50,7 +47,7 @@ class CirclePinDecoration extends PinDecoration implements SupportGap {
     TextStyle errorTextStyle,
     String hintText,
     TextStyle hintTextStyle,
-    Color solidColor,
+    ColorBuilder bgColorBuilder,
   }) {
     return CirclePinDecoration(
       textStyle: textStyle ?? this.textStyle,
@@ -59,24 +56,29 @@ class CirclePinDecoration extends PinDecoration implements SupportGap {
       errorTextStyle: errorTextStyle ?? this.errorTextStyle,
       hintText: hintText ?? this.hintText,
       hintTextStyle: hintTextStyle ?? this.hintTextStyle,
-      solidColor: this.solidColor,
-      strokeColor: this.strokeColor,
+      strokeColorBuilder: this.strokeColorBuilder,
       strokeWidth: this.strokeWidth,
-      enteredColor: this.enteredColor,
       gapSpace: this.gapSpace,
       gapSpaces: this.gapSpaces,
+      bgColorBuilder: this.bgColorBuilder,
     );
   }
 
   @override
-  PinEntryType get pinEntryType => PinEntryType.customized;
+  PinEntryType get pinEntryType => PinEntryType.circle;
+
+  @override
+  void notifyChange(String pin) {
+    strokeColorBuilder.notifyChange(pin);
+    bgColorBuilder?.notifyChange(pin);
+  }
 
   @override
   void drawPin(
     Canvas canvas,
     Size size,
     String text,
-    pinLength,
+    int pinLength,
     ThemeData themeData,
   ) {
     /// Calculate the height of paint area for drawing the pin field.
@@ -92,16 +94,14 @@ class CirclePinDecoration extends PinDecoration implements SupportGap {
     }
 
     Paint borderPaint = Paint()
-      ..color = strokeColor
       ..strokeWidth = strokeWidth
       ..style = PaintingStyle.stroke
       ..isAntiAlias = true;
 
     /// Assign paint if [solidColor] is not null
     Paint insidePaint;
-    if (solidColor != null) {
+    if (bgColorBuilder != null) {
       insidePaint = Paint()
-        ..color = solidColor
         ..style = PaintingStyle.fill
         ..isAntiAlias = true;
     }
@@ -134,21 +134,12 @@ class CirclePinDecoration extends PinDecoration implements SupportGap {
 
     /// Draw the each shape of pin.
     for (int i = 0; i < pinLength; i++) {
-      if (i < text.length && enteredColor != null) {
-        borderPaint.color = enteredColor;
-      } else if (errorText != null && errorText.isNotEmpty) {
+      if (errorText != null && errorText.isNotEmpty) {
         /// only draw error-color as border-color or solid-color
         /// if errorText is not null
-        if (solidColor == null) {
-          borderPaint.color = errorTextStyle.color;
-        } else {
-          insidePaint = Paint()
-            ..color = errorTextStyle.color
-            ..style = PaintingStyle.fill
-            ..isAntiAlias = true;
-        }
+        borderPaint.color = errorTextStyle.color;
       } else {
-        borderPaint.color = strokeColor;
+        borderPaint.color = strokeColorBuilder.indexProperty(i);
       }
       centerPoints[i] = startX + radius;
       canvas.drawCircle(
@@ -157,6 +148,7 @@ class CirclePinDecoration extends PinDecoration implements SupportGap {
         borderPaint,
       );
       if (insidePaint != null) {
+        insidePaint..color = bgColorBuilder.indexProperty(i);
         canvas.drawCircle(
           Offset(startX + radius, startY),
           radius - strokeWidth / 2,
@@ -243,9 +235,8 @@ class CirclePinDecoration extends PinDecoration implements SupportGap {
           strokeWidth == other.strokeWidth &&
           gapSpace == other.gapSpace &&
           gapSpaces == other.gapSpaces &&
-          strokeColor == other.strokeColor &&
-          solidColor == other.solidColor &&
-          enteredColor == other.enteredColor;
+          strokeColorBuilder == other.strokeColorBuilder &&
+          bgColorBuilder == other.bgColorBuilder;
 
   @override
   int get hashCode =>
@@ -253,12 +244,11 @@ class CirclePinDecoration extends PinDecoration implements SupportGap {
       strokeWidth.hashCode ^
       gapSpace.hashCode ^
       gapSpaces.hashCode ^
-      strokeColor.hashCode ^
-      solidColor.hashCode ^
-      enteredColor.hashCode;
+      strokeColorBuilder.hashCode ^
+      bgColorBuilder.hashCode;
 
   @override
   String toString() {
-    return 'CirclePinDecoration{strokeWidth: $strokeWidth, gapSpace: $gapSpace, gapSpaces: $gapSpaces, strokeColor: $strokeColor, solidColor: $solidColor, enteredColor: $enteredColor}';
+    return 'CirclePinDecoration{strokeWidth: $strokeWidth, gapSpace: $gapSpace, gapSpaces: $gapSpaces, strokeColorBuilder: $strokeColorBuilder, bgColorBuilder: $bgColorBuilder}';
   }
 }